move: AddOrEditTemplate under features
Start of the create and edit template workflow refactor
--
The goal is to improve readability, modularity and changeability
of the create and edit template workflow. Each step in the worklow
is conceptually an independent piece that groups UI and related
logic with use-cases together.
--
To start off the refactor of the create and edit, the original files
are only moved from the TemplatesTable's components folder
under newly created features folder.
