+++ src/features/createAndEditTemplate/createTemplate/ui/AddNavigateButton.tsx
