+++ src/features/createAndEditTemplate/defineContent/ui/DefineContentStep.test.tsx
@@ -1,9 +1,9 @@
 import { render } from '@testing-library/react';
-import { useAddTemplateContext } from '../AddTemplateContext';
+import { useAddTemplateContext } from '../../workflow/store/AddTemplateContext';
 import { defaultTemplateItem, testRepositoryParamsResponse } from 'testingHelpers';
 import DefineContentStep from './DefineContentStep';
 
-jest.mock('../AddTemplateContext', () => ({
+jest.mock('../../workflow/store/AddTemplateContext', () => ({
   useAddTemplateContext: jest.fn(),
 }));
