+++ src/features/createAndEditTemplate/reviewTemplateRequest/ui/ReviewStep.test.tsx
@@ -1,10 +1,10 @@
 import { render } from '@testing-library/react';
-import { useAddTemplateContext } from '../AddTemplateContext';
+import { useAddTemplateContext } from '../../workflow/store/AddTemplateContext';
 import { defaultTemplateItem, testRepositoryParamsResponse } from 'testingHelpers';
 import ReviewStep from './ReviewStep';
 import { formatDateDDMMMYYYY } from 'helpers';
 
-jest.mock('../AddTemplateContext', () => ({
+jest.mock('../../workflow/store/AddTemplateContext', () => ({
   useAddTemplateContext: jest.fn(),
 }));
